CW Issue #1384: Move the image registries link to the project validation page
