+++ dev/org.eclipse.codewind.ui/src/org/eclipse/codewind/ui/internal/messages/messages.properties
@@ -220,7 +220,7 @@ SelectProjectPageProjectExistsError=A workspace project already exists with the 
 ProjectValidationPageName=Project Type Detection
 ProjectValidationPageTitle=Confirm Project Type
 ProjectValidationPageDescription=Confirm the detected project type
-ProjectValidationPageMsg=If the detected project type below is correct for project {0}, click Finish. If not, continue to the next page to select the project type.
+ProjectValidationPageMsg=If the detected project type below is correct for project {0}, click Finish. If not, continue to the next page to select the project type. You can also add container registries for your project at this time.
 ProjectValidationPageFailMsg=The project type could not be determined automatically. Continue to the next page to select the project type manually.
 ProjectValidationPageTypeLabel=Type:
 ProjectValidationPageLanguageLabel=Language:
