+++ dev/org.eclipse.codewind.ui/src/org/eclipse/codewind/ui/internal/wizards/ProjectValidationPage.java
@@ -12,26 +12,36 @@
 package org.eclipse.codewind.ui.internal.wizards;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.List;
 
 import org.eclipse.codewind.core.internal.Logger;
 import org.eclipse.codewind.core.internal.cli.ProjectUtil;
 import org.eclipse.codewind.core.internal.connection.CodewindConnection;
+import org.eclipse.codewind.core.internal.connection.ImagePushRegistryInfo;
+import org.eclipse.codewind.core.internal.connection.RegistryInfo;
 import org.eclipse.codewind.core.internal.constants.ProjectInfo;
 import org.eclipse.codewind.ui.internal.IDEUtil;
 import org.eclipse.codewind.ui.internal.messages.Messages;
+import org.eclipse.codewind.ui.internal.prefs.RegistryManagementDialog;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.SubMonitor;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.operation.IRunnableWithProgress;
+import org.eclipse.jface.window.Window;
 import org.eclipse.jface.wizard.WizardPage;
 import org.eclipse.osgi.util.NLS;
 import org.eclipse.swt.SWT;
+import org.eclipse.swt.events.SelectionAdapter;
+import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.Link;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.ui.PlatformUI;
 
@@ -101,6 +111,62 @@ public class ProjectValidationPage extends WizardPage {
 		languageText.setText("");
 		languageText.setLayoutData(new GridData(GridData.FILL, GridData.CENTER, true, false));
 		IDEUtil.normalizeBackground(languageText, composite);
+
+		if (!connection.isLocal()) {
+			// Manage registries link
+			new Label(composite, SWT.NONE).setLayoutData(new GridData(GridData.FILL, GridData.FILL, false, false, 2, 2));
+			Composite manageRegistriesComp = new Composite(composite, SWT.NONE);
+			manageRegistriesComp.setLayout(new GridLayout(1, false));
+			manageRegistriesComp.setLayoutData(new GridData(GridData.BEGINNING, GridData.END, false, false, 2, 1));
+			
+			Link manageRegistriesLink = new Link(manageRegistriesComp, SWT.NONE);
+			manageRegistriesLink.setText(Messages.ManageRegistriesLinkLabel + " <a>" + Messages.ManageRegistriesLinkText + "</a>");
+			if (connection.isLocal()) {
+				manageRegistriesLink.setToolTipText(Messages.ManageRegistriesLinkTooltipLocal);
+			} else {
+				manageRegistriesLink.setToolTipText(Messages.ManageRegistriesLinkTooltip);
+			}
+			manageRegistriesLink.setLayoutData(new GridData(GridData.BEGINNING, GridData.CENTER, false, false));
+	
+			manageRegistriesLink.addSelectionListener(new SelectionAdapter() {
+				@Override
+				public void widgetSelected(SelectionEvent event) {
+					try {
+						List<RegistryInfo> regList = connection.requestRegistryList();
+						ImagePushRegistryInfo pushReg = connection.requestGetPushRegistry();
+						RegistryManagementDialog regDialog = new RegistryManagementDialog(getShell(), connection, regList, pushReg);
+						if (regDialog.open() == Window.OK) {
+							if (regDialog.hasChanges()) {
+								IRunnableWithProgress runnable = new IRunnableWithProgress() {
+									@Override
+									public void run(IProgressMonitor monitor) throws InvocationTargetException {
+										SubMonitor mon = SubMonitor.convert(monitor, Messages.RegUpdateTask, 100);
+										IStatus status = regDialog.updateRegistries(mon.split(75));
+										if (!status.isOK()) {
+											throw new InvocationTargetException(status.getException(), status.getMessage());
+										}
+										if (mon.isCanceled()) {
+											return;
+										}
+									}
+								};
+								try {
+									getWizard().getContainer().run(true, true, runnable);
+								} catch (InvocationTargetException e) {
+									MessageDialog.openError(getShell(), Messages.RegUpdateErrorTitle, e.getMessage());
+									return;
+								} catch (InterruptedException e) {
+									// The user cancelled the operation
+									return;
+								}
+							}
+						}
+					} catch (Exception e) {
+						MessageDialog.openError(getShell(), Messages.RegListErrorTitle, NLS.bind(Messages.RegListErrorMsg, e));
+					}
+				}
+			});
+		}
 		
 		updatePage(false);
